issue#6 - Refractor skywire-update. 'Invalid command' was returned when using skywire-update [config-path]. It turns out that adding cobra subcommands prevents RootCmd from accepting path arguments. Therefore, 'updateCmd' was created so update services can take an optional argument. Update will be based on config path loaded from the following order: 1) CLI argument 2) env 3) list of default path.
