+++ cmd/skywire-updater/commands/gen-config.go
@@ -6,11 +6,10 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/spf13/cobra"
-	"gopkg.in/yaml.v2"
-
 	"github.com/skycoin/skywire-updater/internal/pathutil"
 	"github.com/skycoin/skywire-updater/pkg/update"
+	"github.com/spf13/cobra"
+	"gopkg.in/yaml.v2"
 )
 
 const (
